refactor: updater status text into separate component
Moved the updater status label logic from the splash update page to a new UpdaterStatusText component for better reusability and cleaner code. Updated the splash update page to use the new component.
